+++ tests/data/_.fixes.js
@@ -16,11 +16,13 @@ import $N from './forms/fixes/form.n';
 import $O from './forms/fixes/form.o';
 import $P from './forms/fixes/form.p';
 import $Q from './forms/fixes/form.q';
+import $Q1 from './forms/fixes/form.q1';
+import $Q2 from './forms/fixes/form.q2';
 import $R from './forms/fixes/form.r';
 
 export default {
 
-  $A, $B, $C, $D, $E, $F, $G, $H, $I, $L, $M, $N, $O, $P, $Q, $R,
+  $A, $B, $C, $D, $E, $F, $G, $H, $I, $L, $M, $N, $O, $P, $Q, $Q1, $Q2, $R,
 
 };