+++ tests/data/forms/fixes/form.q2.js
@@ -1,0 +1,19 @@
+import { Form } from '../../../../src';
+
+const fields = [
+  'multiselectArray',
+  'multiselectObject',
+  'tags[]',
+  'tags[].id',
+  'tags[].name',
+];
+
+const values = {
+  multiselectArray: ['iMac', 'iPhone'],
+  multiselectObject: { value: 'watch', label: 'Watch' },
+};
+
+class NewForm extends Form {}
+
+export default new NewForm({ fields, values }, { name: 'Fixes-Q2' });
+