+++ tests/data/forms/nested/form.t.js
@@ -2,6 +2,7 @@ import { Form } from '../../../../src';
 
 const fields = [
   'hobbies[]',
+  'member.data[]',
   'member.hobbies[]',
   'member.info[]',
   'member.info[].firstname',
@@ -18,11 +19,11 @@ class NewForm extends Form {
   hooks() {
     return {
       onInit(form) {
-        form.$('hobbies').add('AAA');
-        form.$('hobbies').add('BBB');
-        form.$('hobbies').add('CCC');
-        form.$('hobbies').add('DDD');
-        form.$('hobbies').add('EEE');
+        form.$('hobbies').add({ value: 'AAA' });
+        form.$('hobbies').add({ value: 'BBB' });
+        form.$('hobbies').add({ value: 'CCC' });
+        form.$('hobbies').add({ value: 'DDD' });
+        form.$('hobbies').add({ value: 'EEE' });
         // hobbies[] length: 5
 
         form.del('hobbies[0]');
@@ -31,12 +32,12 @@ class NewForm extends Form {
         // hobbies[] length should be 2
 
         // form.add('member.hobbies'); // OLD
-        form.$('member.hobbies').add('AAA'); // 1
-        form.$('member.hobbies').add('BBB'); // 2
-        form.$('member.hobbies').add('CCC'); // 3
-        form.$('member.hobbies').add('DDD'); // 4
-        form.$('member.hobbies').add('EEE'); // 5
-        form.$('member.hobbies').add('FFF'); // 6
+        form.$('member.hobbies').add({ value: 'AAA' }); // 1
+        form.$('member.hobbies').add({ value: 'BBB' }); // 2
+        form.$('member.hobbies').add({ value: 'CCC' }); // 3
+        form.$('member.hobbies').add({ value: 'DDD' }); // 4
+        form.$('member.hobbies').add({ value: 'EEE' }); // 5
+        form.$('member.hobbies').add({ value: 'FFF' }); // 6
         // member.hobbies[] length: 6
 
         form.del('member.hobbies[0]');
@@ -45,22 +46,47 @@ class NewForm extends Form {
         form.$('member').del('hobbies[4]');
         // hobbies[] length should be 2
 
-        form.$('member.info').add({
+        // add multiple fileds at once
+        form.$('member.info').add([{
           firstname: 'AAA',
           lastname: 'BBB',
           type: 'XXX',
+        }, {
+          firstname: 'AAA-2',
+          lastname: 'BBB-2',
+          type: 'XXX-2',
+        }]);
+
+        form.$('member.info').add([{
+          firstname: '111',
+          lastname: '222',
+          type: 'YYY',
+        }]);
+
+        form.$('member.data').add([{
+          update: 'update-value',
+        }]);
+
+        form.$('member.data[0]').add({
+          name: 'subzero',
+          value: 'subzero-value',
+          label: 'subzero-label',
         });
 
-        form.$('notIncrementalFields').add('XXX', {
+        form.$('notIncrementalFields').add({
           key: 'notIncrementalKey',
+          value: 'XXX',
+          label: 'XXX-label',
         });
 
-        form.$('notIncrementalFields').add('XXX', {
-          key: 'anotherNotIncrementalKey',
+        form.$('notIncrementalFields').add({
+          name: 'anotherNotIncrementalKey',
+          value: 'XXX',
         });
 
-        form.$('notIncrementalFields').add('XXX', {
-          key: 'anotherOneNotIncrementalKey',
+        form.$('notIncrementalFields').add({
+          name: 'anotherOneNotIncrementalKey',
+          value: 'XXX',
         });
 
         form.del('notIncrementalFields[anotherOneNotIncrementalKey]');